Create load courses Thunk
Modify the course actions to load course information from the API.
Change the CREATE_COURSE action to  LOAD_COURSES_SUCCESS in the action
types. The success suffix on the end of the name is there for clarity
(that the async request was successful and returned all the info from
the API call).