Add author reducer
Add a reducer for author actions, similar to the course reducer. Move
initial state to a separate file and use it within the reducers. This
makes it easier to keep track of default state as the number of
reducers grows. Dispatch loadAuthors in the entry point of the
application (index.js).